+++ kaggle/api/kaggle_api.py
@@ -49,6 +49,184 @@ class KaggleApi(object):
             api_client = ApiClient()
         self.api_client = api_client
 
+    def competition_download_leaderboard(self, id, **kwargs):  # noqa: E501
+        """Download competition leaderboard  # noqa: E501
+
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async=True
+        >>> thread = api.competition_download_leaderboard(id, async=True)
+        >>> result = thread.get()
+
+        :param async bool
+        :param str id: Competition name (required)
+        :return: Result
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+        kwargs['_return_http_data_only'] = True
+        if kwargs.get('async'):
+            return self.competition_download_leaderboard_with_http_info(id, **kwargs)  # noqa: E501
+        else:
+            (data) = self.competition_download_leaderboard_with_http_info(id, **kwargs)  # noqa: E501
+            return data
+
+    def competition_download_leaderboard_with_http_info(self, id, **kwargs):  # noqa: E501
+        """Download competition leaderboard  # noqa: E501
+
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async=True
+        >>> thread = api.competition_download_leaderboard_with_http_info(id, async=True)
+        >>> result = thread.get()
+
+        :param async bool
+        :param str id: Competition name (required)
+        :return: Result
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+
+        all_params = ['id']  # noqa: E501
+        all_params.append('async')
+        all_params.append('_return_http_data_only')
+        all_params.append('_preload_content')
+        all_params.append('_request_timeout')
+
+        params = locals()
+        for key, val in six.iteritems(params['kwargs']):
+            if key not in all_params:
+                raise TypeError(
+                    "Got an unexpected keyword argument '%s'"
+                    " to method competition_download_leaderboard" % key
+                )
+            params[key] = val
+        del params['kwargs']
+        # verify the required parameter 'id' is set
+        if ('id' not in params or
+                params['id'] is None):
+            raise ValueError("Missing the required parameter `id` when calling `competition_download_leaderboard`")  # noqa: E501
+
+        collection_formats = {}
+
+        path_params = {}
+        if 'id' in params:
+            path_params['id'] = params['id']  # noqa: E501
+
+        query_params = []
+
+        header_params = {}
+
+        form_params = []
+        local_var_files = {}
+
+        body_params = None
+        # Authentication setting
+        auth_settings = ['basicAuth']  # noqa: E501
+
+        return self.api_client.call_api(
+            '/competitions/{id}/leaderboard/download', 'GET',
+            path_params,
+            query_params,
+            header_params,
+            body=body_params,
+            post_params=form_params,
+            files=local_var_files,
+            response_type='Result',  # noqa: E501
+            auth_settings=auth_settings,
+            async=params.get('async'),
+            _return_http_data_only=params.get('_return_http_data_only'),
+            _preload_content=params.get('_preload_content', True),
+            _request_timeout=params.get('_request_timeout'),
+            collection_formats=collection_formats)
+
+    def competition_view_leaderboard(self, id, **kwargs):  # noqa: E501
+        """VIew competition leaderboard  # noqa: E501
+
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async=True
+        >>> thread = api.competition_view_leaderboard(id, async=True)
+        >>> result = thread.get()
+
+        :param async bool
+        :param str id: Competition name (required)
+        :return: Result
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+        kwargs['_return_http_data_only'] = True
+        if kwargs.get('async'):
+            return self.competition_view_leaderboard_with_http_info(id, **kwargs)  # noqa: E501
+        else:
+            (data) = self.competition_view_leaderboard_with_http_info(id, **kwargs)  # noqa: E501
+            return data
+
+    def competition_view_leaderboard_with_http_info(self, id, **kwargs):  # noqa: E501
+        """VIew competition leaderboard  # noqa: E501
+
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async=True
+        >>> thread = api.competition_view_leaderboard_with_http_info(id, async=True)
+        >>> result = thread.get()
+
+        :param async bool
+        :param str id: Competition name (required)
+        :return: Result
+                 If the method is called asynchronously,
+                 returns the request thread.
+        """
+
+        all_params = ['id']  # noqa: E501
+        all_params.append('async')
+        all_params.append('_return_http_data_only')
+        all_params.append('_preload_content')
+        all_params.append('_request_timeout')
+
+        params = locals()
+        for key, val in six.iteritems(params['kwargs']):
+            if key not in all_params:
+                raise TypeError(
+                    "Got an unexpected keyword argument '%s'"
+                    " to method competition_view_leaderboard" % key
+                )
+            params[key] = val
+        del params['kwargs']
+        # verify the required parameter 'id' is set
+        if ('id' not in params or
+                params['id'] is None):
+            raise ValueError("Missing the required parameter `id` when calling `competition_view_leaderboard`")  # noqa: E501
+
+        collection_formats = {}
+
+        path_params = {}
+        if 'id' in params:
+            path_params['id'] = params['id']  # noqa: E501
+
+        query_params = []
+
+        header_params = {}
+
+        form_params = []
+        local_var_files = {}
+
+        body_params = None
+        # Authentication setting
+        auth_settings = ['basicAuth']  # noqa: E501
+
+        return self.api_client.call_api(
+            '/competitions/{id}/leaderboard/view', 'GET',
+            path_params,
+            query_params,
+            header_params,
+            body=body_params,
+            post_params=form_params,
+            files=local_var_files,
+            response_type='Result',  # noqa: E501
+            auth_settings=auth_settings,
+            async=params.get('async'),
+            _return_http_data_only=params.get('_return_http_data_only'),
+            _preload_content=params.get('_preload_content', True),
+            _request_timeout=params.get('_request_timeout'),
+            collection_formats=collection_formats)
+
     def competitions_data_download_file(self, id, file_name, **kwargs):  # noqa: E501
         """Download competition data file  # noqa: E501